+++ cas9-ast/pom.xml
@@ -9,13 +9,19 @@
     <relativePath>..</relativePath>
   </parent>
 
-  <artifactId>cas9-parser</artifactId>
-  <name>Pitest Cas9 AST Parser</name>
+  <artifactId>cas9-ast</artifactId>
+  <name>Pitest Cas9 AST</name>
 
   <dependencies>
     <dependency>
       <groupId>com.github.javaparser</groupId>
       <artifactId>javaparser-symbol-solver-core</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+      <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>com.google.code.gson</groupId>
