+++ cas9-engine/src/main/java/org/pitest/mutationtest/engine/cas9/config/Cas9Mutater.java
@@ -1,12 +1,13 @@
-package org.pitest.mutationtest.engine.cas9;
+package org.pitest.mutationtest.engine.cas9.config;
 
 import java.util.Collection;
 import java.util.function.Predicate;
 import org.pitest.classinfo.ClassByteArraySource;
+import org.pitest.mutationtest.engine.cas9.AstGregorMutater;
 import org.pitest.mutationtest.engine.gregor.MethodInfo;
 import org.pitest.mutationtest.engine.gregor.MethodMutatorFactory;
 
-public class Cas9Mutater extends AstParserMutater {
+public class Cas9Mutater extends AstGregorMutater {
 
   public Cas9Mutater(Predicate<MethodInfo> filter, ClassByteArraySource byteSource,
       Collection<MethodMutatorFactory> mutators) {
